chore: extract shared groupStatusByBucket to eliminate bucket-grouping duplication
getCampaign.mjs and rebroadcast/route.js both grouped h1_status rows
by bucket with identical logic. Extracted to shared bucketing utility.
